switch to u64

+++ crates/libafl_qemu/src/modules/usermode/asan_host.rs
@@ -1202,7 +1202,7 @@ where
     if let Some(asan_mappings) = &h.asan_mappings
         && asan_mappings
             .iter()
-            .any(|m| m.start() <= pc && pc < m.end())
+            .any(|m| m.start() <= (pc as u64) && (pc as u64) < m.end())
     {
         return None;
     }
@@ -1298,7 +1298,7 @@ where
     if let Some(asan_mappings) = &h.asan_mappings
         && asan_mappings
             .iter()
-            .any(|m| m.start() <= pc && pc < m.end())
+            .any(|m| m.start() <= (pc as u64) && (pc as u64) < m.end())
     {
         return Some(0);
     }
